Update bootstrap-notify.js
I've had issues with wrong positioning of notifications when closing one notification and adding new one simultaneously. I've been able to track this problem down to the close method which was not adding 'data-closing' attribute and therefore the positioning of new notification was wrong - it was above the old one (which was already closing) and after the old one was closed, it left a a blank space. Using 'attr()' method instead of 'data()' solved the problem (i.e. 'data-closing' attribute is generated as expected).