many fixes
Fixed an issue with exit animations not properly displaying when mixed
(say FadeIn and SlideOut).

Created a function (notifyCloseExcept) that allows closing flashes
except for certain types of flashes (so, you can persist Warnings but
not Infos). Expanded NotiftyClose() to support passing in notification
types to remove only those types of notifications and added the ability
to include arbitrary jquery selectors to target  removing notifications
in more ways.